Tweak another CI dependency version limit
Adding a seemingly unnecessary requirement (because Bundler should
already do this normally) to use the most recent 0.x version of
'cuke_modeler' when testing against the 0.x series. It does, however,
solve the problem of a needlessly old version of the 0.x series from
being used during testing (both locally and in CI).